TASK#15: Creates film reducer fucntion for actions related to film
