+++ project/src/store/data/data-actions.ts
@@ -6,8 +6,6 @@ enum DataActionType {
   SetFilmsLoaded = 'data/set-films-loaded',
   SetPromoFilmLoaded = 'data/set-promo-film-loaded',
   UpdateFilm = 'data/update-film',
-  SetMyListButtonDisabled = 'data/set-my-list-button-disabled',
-  Unknown = 'data/unknown',
 }
 
 const setFilms = (films: Film[]) => ({
@@ -33,18 +31,12 @@ const updateFilm = (film: Film) => ({
   payload: { film },
 } as const);
 
-const setMyListButtonDisabled = (status: boolean) => ({
-  type: DataActionType.SetMyListButtonDisabled,
-  payload: { status },
-} as const);
-
 export {
   setFilms,
   setFilmsLoaded,
   setPromoFilm,
   setPromoFilmLoaded,
   updateFilm,
-  setMyListButtonDisabled,
   DataActionType
 };
 
@@ -53,5 +45,4 @@ export type DataActions =
   | ReturnType<typeof setPromoFilm>
   | ReturnType<typeof setFilmsLoaded>
   | ReturnType<typeof updateFilm>
-  | ReturnType<typeof setMyListButtonDisabled>
   | ReturnType<typeof setPromoFilmLoaded>;
