+++ project/src/store/film/film.test.ts
@@ -1,0 +1,11 @@
+import { setMyListButtonDisabled } from './film-actions';
+import { filmReducer } from './film-reducer';
+
+
+describe('Film reducer', () => {
+  it('should set mylist button disabled', () => {
+    expect( filmReducer( undefined, setMyListButtonDisabled(true))).toEqual({
+      myListButtonDisabled: true,
+    });
+  });
+});
